Construct and use an aura map
Generalize the interdiction field code to work for any hypothetical aura or area-of-effect status field.
Each AI::Step, a map of the current positions of any aura emitter is built for the player's system.
This map can then be used to compute a vector that leads away from all auras a ship may be experiencing.