Rename CanJump to JumpBlocked, apply only to ships which wish to jump
A ship may not be able to jump for any number of reasons, and the response to those reasons should not be the same as the response to being actively blocked from jumping.
Also, if a ship is not trying to jump, it should not care if it is under an enemy's jump interdiction aura.